+++ console/src/components/Common/ConfigureTransformation/PayloadOptionsTransforms.tsx
@@ -48,9 +48,9 @@ const PayloadOptionsTransforms: React.FC<PayloadOptionsTransformsProps> = ({
     <div className="m-md pl-lg pr-sm border-l border-l-gray-400">
       <div className="mb-md">
         <NumberedSidebar
-          number="1"
           title="Sample Input"
           description="Sample input defined by your Action Defintion."
+          number="1"
         >
           <button
             type="button"
@@ -71,10 +71,16 @@ const PayloadOptionsTransforms: React.FC<PayloadOptionsTransformsProps> = ({
 
       <div className="mb-md">
         <NumberedSidebar
-          number="2"
           title="Configure Request Body"
-          description="The template which will transform your request body into the
-          required specification."
+          description={
+            <span>
+              The template which will transform your request body into the
+              required specification. You can use{' '}
+              <code className="text-xs">$body</code> to access the original
+              request body
+            </span>
+          }
+          number="2"
           url="https://hasura.io/docs/latest/graphql/core/actions/transforms.html#request-body"
         />
         <TemplateEditor
@@ -87,10 +93,10 @@ const PayloadOptionsTransforms: React.FC<PayloadOptionsTransformsProps> = ({
 
       <div className="mb-md">
         <NumberedSidebar
-          number="3"
           title="Transformed Request Body"
           description="Sample request body to be delivered based on your input and
           transformation template."
+          number="3"
         >
           {showRequestContentTypeOptions ? (
             <select